Separate unit tests in util.go specific for Windows
