+++ internal/aws/containerinsight/utils_test.go
@@ -5,7 +5,6 @@ package containerinsight
 import (
 	"fmt"
 	"log"
-	"os"
 	"strconv"
 	"strings"
 	"testing"
@@ -866,15 +865,4 @@ func TestConvertToOTLPMetricsForPodContainerStatusMetrics(t *testing.T) {
 	}
 	md = ConvertToOTLPMetrics(fields, tags, zap.NewNop())
 	checkMetricsAreExpected(t, md, fields, tags, expectedUnits)
-}
-
-func TestHostProcessContainer(t *testing.T) {
-	os.Setenv(RunInContainer, "True")
-	assert.Equal(t, IsWindowsHostProcessContainer(), false)
-
-	os.Setenv(RunAsHostProcessContainer, "True")
-	assert.Equal(t, IsWindowsHostProcessContainer(), true)
-
-	os.Unsetenv(RunInContainer)
-	os.Unsetenv(RunAsHostProcessContainer)
 }
